+++ com/eurekaAccounts/stocks/vo/StockFundementalsVO.java
@@ -1,8 +1,10 @@
 package com.eurekaAccounts.stocks.vo;
 
+import java.math.BigDecimal;
+
 public class StockFundementalsVO implements Comparable<StockFundementalsVO> {
     public String tickerSymbol;
-    public int sectorId;
+    public BigDecimal sectorId;
     public int subSectorId;
     public long marketCap;
     public float currentRatio;
@@ -10,7 +12,7 @@ public class StockFundementalsVO implements Comparable<StockFundementalsVO> {
     public StockFundementalsVO() {
     }
 
-    public StockFundementalsVO(String tickerSymbol, int sectorId, int subSectorId, long marketCap, float currentRatio) {
+    public StockFundementalsVO(String tickerSymbol, BigDecimal sectorId, int subSectorId, long marketCap, float currentRatio) {
         this.tickerSymbol = tickerSymbol;
         this.sectorId = sectorId;
         this.subSectorId = subSectorId;
@@ -26,11 +28,11 @@ public class StockFundementalsVO implements Comparable<StockFundementalsVO> {
         this.tickerSymbol = tickerSymbol;
     }
 
-    public int getSectorId() {
+    public BigDecimal getSectorId() {
         return sectorId;
     }
 
-    public void setSectorId(int sectorId) {
+    public void setSectorId(BigDecimal sectorId) {
         this.sectorId = sectorId;
     }
